Add files via upload
Avocado is a fruit consumed by people heavily in the United States. 

Avocado, also called alligator pear, fruit of Persea americana of the family Lauraceae, a tree native to the Western Hemisphere from Mexico south to the Andean regions. Avocado fruits have greenish or yellowish flesh with a buttery consistency and a rich, nutty flavour. They are often eaten in salads, and in many parts of the world they are eaten as a dessert. Mashed avocado is the principal ingredient of guacamole, a characteristic sauce in Mexican cuisine. Avocados provide thiamin, riboflavin, and vitamin A, and in some varieties the flesh contains as much as 25 percent unsaturated oil.